Make the 'log' submodule dep optional
Currently, trying to actually depend on libngrok requires getting two
modules, the main one, and the 'log' one.

Go modules make this have confusing UX!

Right now, this UX is broken! We would have needed to set a more specific version
in the `go.mod` for the log dependency (i.e. v0.0.1 after tagging a release).

This updates how logging is structured to copy the log interface into
both locations, which due to go's structural typing should work for our
needs, but also importantly makes pulling in this code as a dependency
easier and less confusing.

To quote the rob pike proverb: "A little copying is better than a little
dependency."