log: add licenses to submodules
Currently, if you go to
https://pkg.go.dev/github.com/ngrok/libngrok-go/log, you get
"Documentation not displayed due to license restrictions."

I'm hopeful that this will fix it.

I've tested this by using `go run golang.org/x/pkgsite/cmd/pkgsite` to
run it locally and verified that the 'log' submodule has the little
'Redistributable License' tickmark on it with this change.