session: allow ReconnectingSession to retry temporary failures
In the case of a temporary failure in ReconnectingSession, an error is
placed into the stateChanges channel. This error does not necessarily
mean that failure is permanent and that the connection process should
abort. For example, in the case of a dial error when initially
connecting the session, we desire to retry dialing with backoff as
implemented by ReconnectingSession, unless the library consumer
intentionally wants to cancel connecting based on that error.

Only in the case of a permanent connection failure do we want to cancel
the connection process and return an error to the consumer.

Therefore, we expose all connection errors to the library consumer via
the Disconnect handler, allowing them, if they so choose, to .Close() an
open session that is erroring (e.g., on authentication or restarting
tunnel binds). The Connect function returns only upon permanent failure,
in which case it returns nil and the last error read from stateChanges;
or upon connection success, in which case it returns a happy Session. In
the case of permanent failure, the Disconnect handler will have already
received the error; thus, for consumers defining a Disconnect handler it
is not necessary to handle the error returned from Connect.

In addition, to allow consumers to close a not-yet-connected session, we
make sure to always pass a well-formed Session to the Disconnect
handler.

Finally, we fix a situation in which canceling the context leaves the
underlying session running.