Update muxado to fix a heartbeat-timeout related edge-case
The upstream muxado change is https://github.com/ngrok/muxado-go/pull/2

In practice, this seems to be necessary on windows, where time is a
little weird.

This does the straightforward plumbing to update this repo.

I don't believe this counts as a breaking change here since I'm pretty
sure none of this leaks into the public user interface of the package.