feat: Increase logging around proxy conn
I was looking specifically for the client.addr and didn't see it being
logged anywhere. We were previously logging this in the ngrok-operator
project when we were handling dialing the upstream and joining
connections on our own. Now that we are using an endpoint forwarder to
clean up code in the ngrok-operator, we no longer see these logs
messages when increasing log verbosity.